backend/requirements.txt
# FastAPI Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
pymongo==4.8.0
motor==3.5.0

# ML / Deep Learning
tensorflow==2.15.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2

# Explainable AI
shap==0.45.0

# WebSocket
websockets==12.0

# Utilities
python-dotenv==1.0.1
